Move length, modify_many_values, modify_values to sqlite.query_helpers
